Added abstract class WrappedSamplesQuery
